[refactor] issue-#3 #4 #5 : GameService 일부 수정 및 Dto 형태 개편

+++ src/main/java/likelion/hufsglobal/lgtu/runwithmate/domain/gameroom/RoomUpdateReqDto.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class RoomUpdateReqDto {
     private Long betPoint;
-    private Integer timeLimit;
+    private Long timeLimit;
 }
